+++ src/find_dependencies/dependency_finder.py
@@ -19,36 +19,49 @@ class DependencyFinder:
         self.test_number = 0
 
     def find_dependencies(self):
-        failed1 = self.run_tests(self.session.items)
+        reversed_first = self.session.config.getoption("reversed_first")
         reversed_items = self.session.items[::-1]
-        failed2 = self.run_tests(reversed_items)
-        # tests that fail in both runs are not considered
+        if reversed_first:
+            items1, items2 = reversed_items, self.session.items
+        else:
+            items1, items2 = self.session.items, reversed_items
+        failed1 = self.run_tests(items1)
+        failed2 = self.run_tests(items2)
+        always_failing_items = failed1.intersection(failed2)
+
+        # tests failing in both runs are not considered
         failed1, failed2 = failed1 - failed2, failed2 - failed1
         for item in sorted(failed1, key=str):
-            self.check_failed_item(item, self.session.items)
+            self.check_failed_item(item, items1)
 
         # tests failing in the second run have to checked if they fail
         # permanently afterwards - in this case the we don't try to
         # find the dependency
         for item in sorted(failed2, key=str):
-            self.check_failed_item(item, reversed_items, check_permanent=True)
+            self.check_failed_item(item, items2, check_permanent=True)
 
-        print("\n=================================================")
-        print(f"Run dependency analysis for {len(self.session.items)} tests.")
+        print("\n=====================================================")
+        print(f"Run dependency analysis for {len(items1)} tests.")
         print(f"Executed {self.test_number} tests in "
               f"{self.test_runs} test runs.")
+        if always_failing_items:
+            print("The following tests are always failing and "
+                  "are excluded from the analysis:")
+            for item in sorted(always_failing_items):
+                print(f"  {item.nodeid}")
+
         if not self.dependent_items and not self.permanently_failed_items:
             print("No dependent tests found.")
         else:
             if self.permanently_failed_items:
                 print("Tests failing permanently after all tests have run:")
                 for item in self.permanently_failed_items:
-                    print(item.nodeid)
+                    print(f"  {item.nodeid}")
             if self.dependent_items:
                 print("Dependent tests:")
                 for item, dependent in self.dependent_items.items():
-                    print(f"{item.nodeid} depends on {dependent.nodeid}")
-        print("=================================================")
+                    print(f"  {item.nodeid} depends on {dependent.nodeid}")
+        print("=====================================================")
 
     def check_failed_item(self, item, items, failed=True,
                           check_permanent=False):