+++ tests/test_find_dependencies.py
@@ -75,6 +75,9 @@ def test_no_dependencies(test_path):
 
     result = test_path.runpytest("--find-dependencies")
     result.stdout.fnmatch_lines([
+        "The following tests are always failing and are "
+        "excluded from the analysis:",
+        "  test_one.py::test_b",
         "No dependent tests found."
     ])
 
@@ -115,7 +118,7 @@ def test_single_dependency_last_index(test_path):
         "Run dependency analysis for 3 tests.",
         "Executed 7 tests in 3 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_c"
+        "  test_one.py::test_b depends on test_one.py::test_c"
     ])
 
 
@@ -134,7 +137,7 @@ def test_single_dependency_first_index(test_path):
         "Run dependency analysis for 3 tests.",
         "Executed 6 tests in 2 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_a"
+        "  test_one.py::test_b depends on test_one.py::test_a"
     ])
 
 
@@ -154,7 +157,28 @@ def test_single_dependency1(test_path):
         "Run dependency analysis for 4 tests.",
         "Executed 11 tests in 4 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_c"
+        "  test_one.py::test_b depends on test_one.py::test_c"
+    ])
+
+
+def test_single_reversed_first(test_path):
+    test_path.makepyfile(
+        test_one="""
+        flag = True
+        def test_a(): pass
+        def test_b(): assert flag
+        def test_c(): global flag; flag = False
+        def test_d(): pass
+        """
+    )
+
+    result = test_path.runpytest("--find-dependencies", "--reversed-first",
+                                 "-p", "no:randomly")
+    result.stdout.fnmatch_lines([
+        "Run dependency analysis for 4 tests.",
+        "Executed 10 tests in 3 test runs.",
+        "Dependent tests:",
+        "  test_one.py::test_b depends on test_one.py::test_c"
     ])
 
 
@@ -179,7 +203,7 @@ def test_single_dependency2(test_path):
         "Run dependency analysis for 9 tests.",
         "Executed 27 tests in 6 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_g"
+        "  test_one.py::test_b depends on test_one.py::test_g"
     ])
 
 
@@ -202,7 +226,7 @@ def test_single_dependency3(test_path):
         "Run dependency analysis for 7 tests.",
         "Executed 19 tests in 4 test runs.",
         "Dependent tests:",
-        "test_one.py::test_e depends on test_one.py::test_b"
+        "  test_one.py::test_e depends on test_one.py::test_b"
     ])
 
 
@@ -221,7 +245,7 @@ def test_single_dependency1_with_randomly(test_path):
     result.stdout.fnmatch_lines([
         "Run dependency analysis for 4 tests.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_c"
+        "  test_one.py::test_b depends on test_one.py::test_c"
     ])
 
 
@@ -245,7 +269,7 @@ def test_single_dependency2_with_randomly(test_path):
     result.stdout.fnmatch_lines([
         "Run dependency analysis for 9 tests.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_g"
+        "  test_one.py::test_b depends on test_one.py::test_g"
     ])
 
 
@@ -267,7 +291,7 @@ def test_single_dependency3_with_randomly(test_path):
     result.stdout.fnmatch_lines([
         "Run dependency analysis for 7 tests.",
         "Dependent tests:",
-        "test_one.py::test_e depends on test_one.py::test_b"
+        "  test_one.py::test_e depends on test_one.py::test_b"
     ])
 
 
@@ -289,8 +313,8 @@ def test_two_dependencies(test_path):
         "Run dependency analysis for 6 tests.",
         "Executed 21 tests in 7 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_e",
-        "test_one.py::test_d depends on test_one.py::test_e"
+        "  test_one.py::test_b depends on test_one.py::test_e",
+        "  test_one.py::test_d depends on test_one.py::test_e"
     ])
 
 
@@ -321,7 +345,7 @@ def test_single_dependency_in_other_module1(test_path):
         "Run dependency analysis for 7 tests.",
         "Executed 19 tests in 4 test runs.",
         "Dependent tests:",
-        "test_one.py::test_e depends on test_one.py::test_b"
+        "  test_one.py::test_e depends on test_one.py::test_b"
     ])
 
 
@@ -351,7 +375,7 @@ def test_single_dependency_in_other_module2(test_path):
         "Run dependency analysis for 7 tests.",
         "Executed 20 tests in 5 test runs.",
         "Dependent tests:",
-        "test_one.py::test_b depends on test_one.py::test_e"
+        "  test_one.py::test_b depends on test_one.py::test_e"
     ])
 
 
@@ -382,5 +406,39 @@ def test_permanent_dependency(test_path):
         "Run dependency analysis for 5 tests.",
         "Executed 11 tests in 3 test runs.",
         "Tests failing permanently after all tests have run:",
-        "test_one.py::test_b"
+        "  test_one.py::test_b"
+    ])
+
+
+def test_permanent_dependency_reversed_first(test_path):
+    test_path.makepyfile(
+        test_one="""
+        import util
+        def test_a(): pass
+        def test_b(): assert not util.lock_exists()
+        def test_c(): pass
+        def test_d(): util.create_lock()
+        def test_e(): pass
+        """
+    )
+    test_path.makepyfile(
+        util="""
+        import os
+        def create_lock():
+            with open("lock1.lck", "w") as f:
+                f.write("test")
+
+        def lock_exists():
+            return os.path.exists("lock1.lck")
+        """
+    )
+    result = test_path.runpytest("--find-dependencies", "--reversed-first",
+                                 "-p", "no:randomly")
+    result.stdout.fnmatch_lines([
+        "Run dependency analysis for 5 tests.",
+        "Executed 10 tests in 2 test runs.",
+        "The following tests are always failing and are "
+        "excluded from the analysis:",
+        "  test_one.py::test_b",
+        "No dependent tests found."
     ])